refactor(providers): reloca o LinkedIn (background) para src/providers/linkedin/
processLinkedInCapture + findLinkedInPublicationByUrn + build/compute do LinkedIn saem
do controller para o módulo coeso, expostos via linkedinProvider. Corrige o erro de
tipo do destructure engagers (usa ...post direto, byte-compat). Golden-master verde,
27 testes, typecheck 4->3 erros.

Refs #5 #6
